biggest push of the century
I added the scraping tools
Currently using node.js and trying to work with the server-side part, but idk yet
The scrape.js is where all the magic happens
you need to install node.js, and run node scrape.js in the directory to start the server
ISSUE: requirejs is not working yet, and so i can't really test the functionality yet. someone help?